fix(composition): update compositions in edit session
